% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{add_description}
\usage{
add_description(input, label)
}
\arguments{
\item{input}{A vector of stats_units to be added or an AnalysR env}

\item{label}{Label to write in description table}
}
\description{
The add_description function adds a description to the description table.
}
