% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stat_units.R
\name{add_stat_units}
\alias{add_stat_units}
\title{Add stat units}
\usage{
add_stat_units(ids, model = analysr_env)
}
\arguments{
\item{ids}{A vector containing ids of stat units to add.}

\item{model}{An AnalysR env.
Default: \code{analysr_env}}
}
\description{
Add to stat_units table
}
\examples{
add_stat_units(c(144545, 78787, 845))
}
