% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{after}
\alias{after}
\title{after}
\usage{
after(model, condition)
}
\arguments{
\item{model}{An AnalysR model}

\item{condition}{A condition}
}
\description{
The after function is used to return the set of events or measurements that
verify a certain condition observed after an event / measurement.
}
