% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/before.R
\name{before}
\alias{before}
\title{before}
\usage{
before(model, condition)
}
\arguments{
\item{model}{An AnalysR model}

\item{condition}{A condition}
}
\description{
The before function is used to return the set of events or measurements that
verify a certain condition observed before an event / measurement.
}
