% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature.R
\name{create_feature}
\alias{create_feature}
\title{create_feature}
\usage{
create_feature(
  model,
  tag_to_create,
  wanted_tag,
  start,
  end,
  aggregation_method = mean_aggregate
)
}
\arguments{
\item{model}{An AnalysR model}

\item{tag_to_create}{Label to write in description table}

\item{wanted_tag}{The name of the measures we want to work on}

\item{start}{A date marking the beginning of the studied time period.}

\item{end}{A date marking the end of the studied time period.}

\item{aggregation_method}{A function to aggregate data.
Default: mean_aggregate}
}
\description{
create_feature
}
