% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_raw.R
\name{export_raw}
\alias{export_raw}
\title{export_raw}
\usage{
export_raw(model, tag)
}
\arguments{
\item{model}{An AnalysR model}

\item{tag}{A tag}
}
\description{
The export_raw function is an experimental feature.
There is no unit test. This allows you to export data grouped by stat_units.
}
