% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature.R
\name{extract_feature}
\alias{extract_feature}
\title{extract_feature}
\usage{
extract_feature(model, tag)
}
\arguments{
\item{model}{An AnalysR model}

\item{tag}{A tag existing in the description table}
}
\description{
The extract_feature function is quite the same goal as create_feature.
This function create a new coll named with the description tag in
the stat_units table.
The function will extract feature form the description table.
}
