% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_descriptions.R
\name{fill_descriptions}
\alias{fill_descriptions}
\title{Fill descriptions}
\usage{
fill_descriptions(hash, types, data, n = length(hash), model = analysr_env)
}
\arguments{
\item{hash}{a vector containing hash to associate}

\item{types}{a vector of strings (the name of the columns to import)}

\item{data}{the data frame you want to extract data from}

\item{n}{(optional) length of hash if already calculated}

\item{model}{An AnalysR env.
Default: \code{analysr_env}}
}
\description{
A function to fill descriptions table when importing data
(typically those which are not in mandatory)
}
