% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_granularity.R
\name{fix_granularity}
\alias{fix_granularity}
\title{fix_granularity}
\usage{
fix_granularity(
  tag_wanted,
  period_start,
  period_end,
  temporal_granularity,
  stat_unit_wanted = NULL,
  aggregation_method = mean_aggregate,
  impute_method = linear_impute,
  information_lost_after = 5 * temporal_granularity
)
}
\arguments{
\item{tag_wanted}{A string containing the tag of the wanted data.}

\item{period_start}{A date marking the beginning of the studied time period.}

\item{period_end}{A date marking the end of the studied time period.}

\item{temporal_granularity}{A duration fixing the wanted granularity.}

\item{stat_unit_wanted}{An vector of integers containing the stat_units
of the wanted data.
Default: NULL (means every stat_unit)}

\item{aggregation_method}{A function to aggregate data.
Default: mean_aggregate}

\item{impute_method}{A function to impute data.
Default: linear_impute}

\item{information_lost_after}{A duration after which, if no data are found,
no more will be imputed.
Default: 5 * temporal_granularity}
}
\value{
The part of analysr_env$measures that has been modified.
}
\description{
Sets a given granularity to a set of data by aggregating or imputing them.
}
