% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{get_hash}
\alias{get_hash}
\title{Get a hash}
\usage{
get_hash(n)
}
\arguments{
\item{n}{A number.}
}
\value{
A number or verctor depending on \code{n} value
}
\description{
Return a number or verctor depending on \code{n} value
}
\examples{
get_hash(10)
}
