% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_events_csv.R
\name{import_events_csv}
\alias{import_events_csv}
\title{import_events_csv}
\usage{
import_events_csv(
  csv_path,
  stat_unit = "stat_unit",
  date = "date",
  tag = "tag",
  optional_data,
  date_format_func = (function(x) lubridate::parse_date_time(x, date_format_reg)),
  date_format_reg = "ymd-HMS"
)
}
\arguments{
\item{csv_path}{A path to the csv file.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{date}{A string containing the date label.}

\item{tag}{A string containing the tag label.}

\item{optional_data}{A vector containing label to import in descriptions
table.}

\item{date_format_func}{A function to format date with (not required).
Default: \code{lubridate::parse_date_time(x, date_format_reg)}}

\item{date_format_reg}{A expression to format date with (not required).
Default: \code{"ymd-HMS"}}
}
\value{
The events data frame resulted from the merge of imported data
and already imported data
}
\description{
Import events from a CSV file
}
