% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_FHIR.R
\name{import_fhir}
\alias{import_fhir}
\title{import_fhir}
\usage{
import_fhir(
  folder_path,
  date = c("date"),
  begin = c("begin"),
  end = c("end"),
  tag = c("tag"),
  value = c("value")
)
}
\arguments{
\item{folder_path}{A path to the FHIR file.}

\item{date}{A vector of labels containing dates.}

\item{begin}{A vector of labels containing begin dates.}

\item{end}{A vector of labels containing end dates.}

\item{tag}{A vector of labels containing tags.}

\item{value}{A vector of labels containing values.}
}
\value{
Merge of imported data and already imported data
}
\description{
Import data from a FHIR folder
}
