% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_measures_df.R
\name{import_measures_df}
\alias{import_measures_df}
\title{import_measures_df}
\usage{
import_measures_df(
  input,
  stat_unit = "stat_unit",
  date = "date",
  tag = "tag",
  value = "value",
  optional_data,
  status = "status",
  model = analysr_env
)
}
\arguments{
\item{input}{A data frame.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{date}{A string containing the date label.}

\item{tag}{A string containing the tag label.}

\item{value}{A string containing the value label.}

\item{optional_data}{A vector containing label to import in descriptions
table (not required).}

\item{status}{A string containing the status label.}

\item{model}{An AnalysR env.
Default: \code{analysr_env}}
}
\value{
A boolean (\code{TRUE} if no errors)
}
\description{
Import measures from a data frame.
}
