% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_periods_df.R
\name{import_periods_df}
\alias{import_periods_df}
\title{import_periods_df}
\usage{
import_periods_df(
  input,
  stat_unit = "stat_unit",
  begin = "begin",
  end = "end",
  tag = "tag",
  optional_data,
  model = analysr_env
)
}
\arguments{
\item{input}{A data frame.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{begin}{A string containing the begin date label.}

\item{end}{A string containing the end date label.}

\item{tag}{A string containing the tag label.}

\item{optional_data}{A vector containing label to import in descriptions
table.}

\item{model}{An AnalysR env.
Default: \code{analysr_env}}
}
\value{
A boolean (\code{TRUE} if no errors)
}
\description{
Import periods from a data frame.
}
