% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stat_units_csv.R
\name{import_stat_units_csv}
\alias{import_stat_units_csv}
\title{import_stat_units_csv}
\usage{
import_stat_units_csv(
  csv_path,
  stat_unit = "stat_unit",
  optional_data,
  delim = ","
)
}
\arguments{
\item{csv_path}{A path to the csv file.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{optional_data}{A vector containing label to import in descriptions
table.}

\item{delim}{The separator to read csv (not required).
Default: \verb{,}}
}
\value{
TRUE if success
}
\description{
Import stat_units from a CSV file
}
