% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stat_units_df.R
\name{import_stat_units_df}
\alias{import_stat_units_df}
\title{import_stat_units_df}
\usage{
import_stat_units_df(
  input,
  stat_unit = "stat_unit",
  optional_data,
  model = analysr_env
)
}
\arguments{
\item{input}{A data frame.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{optional_data}{A vector containing label to import in descriptions
table.}

\item{model}{An AnalysR env.
Default: \code{analysr_env}}
}
\value{
A boolean (\code{TRUE} if no errors)
}
\description{
Import stat_units from a dataframe.
}
