% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_event.R
\name{induce_event}
\alias{induce_event}
\title{induce_event}
\usage{
induce_event(model = analysr_env, condition, tag_to_create)
}
\arguments{
\item{model}{An AnalysR env.
Default: \code{analysr_env}}

\item{condition}{A condition.}

\item{tag_to_create}{Label to write in events table.}
}
\value{
A AnalysR model
}
\description{
Create event from measures table depending on a condition,
and add them to events table.
}
