% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_events.R
\name{induce_events}
\alias{induce_events}
\title{induce_event}
\usage{
induce_events(condition, event)
}
\arguments{
\item{condition}{A condition.}

\item{event}{An event name.}
}
\value{
The part of event tablethat have been created.
}
\description{
Create event from measures table depending on a condition,
and add them to events table
}
