% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_measure.R
\name{induce_measure}
\alias{induce_measure}
\title{induce_measure}
\usage{
induce_measure(model = analysr_env, tag_to_create, calcul, tag_ref)
}
\arguments{
\item{model}{An AnalysR env}

\item{tag_to_create}{Label to write in measures table}

\item{calcul}{A function containing at least a measures tag}

\item{tag_ref}{A tag that is considered to iter on measures table}
}
\value{
A AnalysR model
}
\description{
induce_measure
}
