% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_period.R
\name{induce_period}
\alias{induce_period}
\title{induce_period}
\usage{
induce_period(model = analysr_env, condition, tag_to_create, duration)
}
\arguments{
\item{condition}{A condition (to be defined).}

\item{tag_to_create}{Label to write in periods table.}

\item{duration}{A duration.}
}
\value{
A AnalysR model
}
\description{
Create period from measures table depending on a condition (on measures or
events),
and add them to periods table
}
