% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_outside.R
\name{inside}
\alias{inside}
\alias{during}
\title{inside}
\usage{
inside(model, period_wanted)

during(model, period_wanted)
}
\arguments{
\item{model}{An AnalysR model.}

\item{period_wanted}{A period from the period table.}
}
\description{
inside

during
}
