% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_env_csv.R
\name{load_env_csv}
\alias{load_env_csv}
\title{Load env csv}
\usage{
load_env_csv(save_path)
}
\arguments{
\item{save_path}{A path to the save folder}
}
\description{
Load saved environment from a folder containing
}
