% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_env_rdata.R
\name{load_env_rdata}
\alias{load_env_rdata}
\title{Load env RData}
\usage{

load_env_rdata(save_path)

}
\arguments{
\item{save_path}{A path to the save file}
}
\description{
Load saved environment from a Rdata file
}
