% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.R
\name{observed}
\alias{observed}
\title{observed}
\usage{
observed(model, condition)
}
\arguments{
\item{model}{An AnalysR model}

\item{condition}{A condition}
}
\description{
This function can be complemented by various keywords that do not work alone,
see the secondary keywords on the documentation.
}
