% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_outside.R
\name{outside}
\alias{outside}
\title{outside}
\usage{
outside(model, period_wanted)
}
\arguments{
\item{model}{An AnalysR model.}

\item{period_wanted}{A period from the period table.}
}
\description{
outside
}
