% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_env_csv.R
\name{save_env_csv}
\alias{save_env_csv}
\title{Save env csv}
\usage{
save_env_csv(save_path)
}
\arguments{
\item{save_path}{A path to save current environment
(existence is not required)}
}
\description{
Save current environment in a folder containing CSV
}
\examples{

save_env_csv("~/")

}
