% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_env_rdata.R
\name{save_env_rdata}
\alias{save_env_rdata}
\title{Save env RData}
\usage{
save_env_rdata(save_path, model = analysr_env)
}
\arguments{
\item{save_path}{A path to save current environment
(existence is required)}

\item{model}{a AnalysR Model (default : \code{analysr_env})}
}
\description{
Save current environment in a Rdata file.
}
\examples{
save_env_rdata("~/")

}
