% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to.R
\name{to}
\alias{to}
\title{to}
\usage{
to(
  model,
  date,
  date_format_func = (function(x) lubridate::parse_date_time(x, date_format_reg)),
  date_format_reg = "ymd-HMS"
)
}
\arguments{
\item{model}{A Analysr model instance.}

\item{date}{A date object or a string to be converted as a date.}

\item{date_format_func}{A function to format date with (not required).
Default: \code{lubridate::parse_date_time(x, date_format_reg)}}

\item{date_format_reg}{A expression to format date with (not required).
Default: \code{"ymd-HMS"}}
}
\description{
to
}
