% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_calc.R
\name{vector_calc}
\alias{vector_calc}
\title{vector_calc}
\usage{
vector_calc(model, tag_to_create, func_to_apply, wanted_tag, start, end)
}
\arguments{
\item{model}{An AnalysR model}

\item{tag_to_create}{Label to write in description table}

\item{func_to_apply}{A function to apply to the vector selected.}

\item{wanted_tag}{The name of the measures we want to work on.
(not mandatory)}

\item{start}{A date marking the beginning of the studied time period
(not mandatory)}

\item{end}{A date marking the end of the studied time period.
(not mandatory)}
}
\description{
This function allow you to calcul data from a simple selection to add a
description. The \code{vector_calc} name can be missleading.
}
\details{
Setting a value to \code{start}, \code{end}, and \code{wanted_tag} would make the
\code{vector_calc} function on measure table, almost like \code{create_feature}.

If you do not provide a start/end, \code{vector_calc} would try to work on the
selection table.

The result is always in the description table.
}
