% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_is.R
\name{who_is}
\alias{who_is}
\title{who_is}
\usage{
who_is(model, condition)
}
\arguments{
\item{model}{An AnalysR model or a list of stat_unit id}

\item{condition}{A condition.}
}
\description{
A function which does the same as \code{described_by} function.
The who_is function allows you to filter the selection table
(or if you want this allows you to refine the request) only on description
of the stat units which is mainly.
}
