% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{add_description}
\usage{
add_description(stat_units, label)
}
\arguments{
\item{stat_units}{A vector of stats_units to be added}

\item{label}{Label to write in description table}
}
\description{
add_description
}
