% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/described_by.R
\name{described_by}
\alias{described_by}
\title{described_by}
\usage{
described_by(model, condition)
}
\arguments{
\item{model}{An AnalysR model}

\item{condition}{A condition.}
}
\description{
The described_by function allows you to filter the selection table
(or if you want this allows you to refine the request).
}
