% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_measures_csv.R
\name{import_measures_csv}
\alias{import_measures_csv}
\title{import_measures_csv}
\usage{
import_measures_csv(
  csv_path,
  stat_unit = "stat_unit",
  date = "date",
  tag = "tag",
  value = "value",
  optional_data,
  status = "status",
  date_format_func = lubridate::ymd_hms,
  date_format_reg
)
}
\arguments{
\item{csv_path}{A path to the csv file.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{date}{A string containing the date label.}

\item{tag}{A string containing the tag label.}

\item{value}{A string containing the value label.}

\item{optional_data}{A vector containing label to import in descriptions
table (not required).}

\item{status}{A string containing the status label.}

\item{date_format_func}{A function to format date with (not required).}

\item{date_format_reg}{A expression to format date with (not required).}
}
\value{
The measures data frame resulted from the merge of imported data
and already imported data
}
\description{
Import measures from a CSV file
}
