% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_measures_csv.R
\name{import_measures_csv}
\alias{import_measures_csv}
\title{import_measures_csv}
\usage{
import_measures_csv(
  csv_path,
  stat_unit = "stat_unit",
  date = "date",
  tag = "tag",
  value = "value"
)
}
\arguments{
\item{csv_path}{A path to the csv file.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{date}{A string containing the date label.}

\item{tag}{A string containing the tag label.}

\item{value}{A string containing the value label.}
}
\value{
The measures data frame resulted from the merge of imported data
and already imported data
}
\description{
Import measures from a CSV file
}
\examples{
import_measures_csv(csv_path, stat_unit, date, tag, value)
}
