% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_periods_csv.R
\name{import_periods_csv}
\alias{import_periods_csv}
\title{import_periods_csv}
\usage{
import_periods_csv(
  csv_path,
  stat_unit = "stat_unit",
  begin = "begin",
  end = "end",
  tag = "tag",
  optional_data,
  date_format_func = (function(x) lubridate::parse_date_time(x, date_format_reg)),
  date_format_reg = "ymd-HMS",
  force_date_format = FALSE,
  delim = ",",
  model = analysr_env
)
}
\arguments{
\item{csv_path}{A path to the csv file.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{begin}{A string containing the begin date label.}

\item{end}{A string containing the end date label.}

\item{tag}{A string containing the tag label.}

\item{optional_data}{A vector containing label to import in descriptions
table.}

\item{date_format_func}{A function to format date with (not required).
Default: \code{lubridate::parse_date_time(x, date_format_reg)}
If you want to use milliseconds \href{https://bit.ly/33JGr6s}{look at this}.}

\item{date_format_reg}{A expression to format date with (not required).
Default: \code{"ymd-HMS"}
For more details see \href{https://bit.ly/3bp3FD0}{this documentation}.}

\item{force_date_format}{Boolean to force date format func (not required).
Default: \code{FALSE}}

\item{delim}{The separator to read csv (not required).
Default: \verb{,}}

\item{model}{An AnalysR env.
Default: \code{analysr_env}}
}
\value{
A boolean (\code{TRUE} if no errors)
}
\description{
Import periods from a CSV file.
}
