% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_periods_csv.R
\name{import_periods_csv}
\alias{import_periods_csv}
\title{import_periods_csv}
\usage{
import_periods_csv(
  csv_path,
  stat_unit = "stat_unit",
  begin = "begin",
  end = "end",
  desc = "desc"
)
}
\arguments{
\item{csv_path}{A path to the csv file.}

\item{stat_unit}{A string containing the stat_unit label.}

\item{begin}{A string containing the begin date label.}

\item{end}{A string containing the end date label.}

\item{desc}{A string containing the desc label.}
}
\value{
The periods data frame resulted from the merge of imported data
and already imported data
}
\description{
Import periods from a CSV file
}
\examples{
import_periods_csv(csv_path, stat_unit, begin, end, desc)
}
