% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_events.R
\name{induce_events}
\alias{induce_events}
\title{induce_event}
\usage{
induce_events(condition, event)
}
\arguments{
\item{condition}{A condition (to be defined).}

\item{duration}{A duration.}
}
\value{
The part of  that have been created.
}
\description{
Create event from measures table depending on a condition,
and add them to events table
}
