% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_period.R
\name{induce_period}
\alias{induce_period}
\title{induce_period}
\usage{
induce_period(condition, tag_to_create, duration)
}
\arguments{
\item{condition}{A condition (to be defined).}

\item{tag_to_create}{A string for the tag to add.}

\item{duration}{A duration.}
}
\value{
The part of  that have been created.
}
\description{
Create period from measures table depending on a condition,
and add them to periods table
}
